% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPI_calculator.R
\name{CPI_calculator}
\alias{CPI_calculator}
\title{CPI_calculator.}
\usage{
CPI_calculator(
  x,
  y = NULL,
  genename,
  genenamey = NULL,
  genestart = 2,
  genesel = NULL,
  onecolumn = FALSE,
  sexcolumn = NULL,
  threshold = 0.9999
)
}
\arguments{
\item{x}{A matrix or data frame containing individual names and gene
information.}

\item{y}{Another gene information matrix or data frame, default is
\code{NULL}.}

\item{genename}{A vector or matrix containing only all gene names.}

\item{genenamey}{When parameter y exists, the name of all genes in matrix y
should also be specified, default is \code{NULL}.}

\item{genestart}{A numeric representing the first column of gene information,
default is 2.}

\item{genesel}{A vector or matrix containing selected gene names, default is
\code{NULL}.}

\item{onecolumn}{A logical value, if \code{TRUE}, one column represents a
gene site, default is \code{FALSE}.}

\item{sexcolumn}{A numeric value representing the gender column, default is
\code{NULL}.}

\item{threshold}{A numeric value representing the threshold of probability of
paternity, default is 0.9999}
}
\value{
A list contains CPI matrix, POP matrix, and parent-child duos that
  identified as true.
}
\description{
Calculate the CPI and POP between individuals.
}
\details{
Calculate the CPI(combined paternity index) and POP(probability of
  paternity) between every two individuals in a matrix or two individuals
  from two matrices.
}
