% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forensic_calculator.R
\name{Forensic_calculator}
\alias{Forensic_calculator}
\title{Forensic_calculator.}
\usage{
Forensic_calculator(
  x,
  y = NULL,
  genename,
  genenamey = NULL,
  genestart = 2,
  genesel = NULL,
  onecolumn = FALSE
)
}
\arguments{
\item{x}{A matrix or data frame containing individual names and gene
information.}

\item{y}{Another gene information matrix or data frame, default is
\code{NULL}.}

\item{genename}{A vector or matrix containing only all gene names.}

\item{genenamey}{When parameter y exists, the name of all genes in matrix y
should also be specified, default is \code{NULL}.}

\item{genestart}{A numeric representing the first column of gene information,
default is 2.}

\item{genesel}{A vector or matrix containing selected gene names, default is
\code{NULL}.}

\item{onecolumn}{A logical value, if \code{TRUE}, one column represents a
gene site, default is \code{FALSE}.}
}
\value{
A matrix contains site names and 7 forensic parameters.
}
\description{
Calculate the forensic parameters for every site in the group.
}
\details{
Forensic parameters including MAF:minor allele frequency, H_obs:
  observed heterozygosity, H_exp: expected heterozygosity, Ae: effective
  number of allele, DP: discrimination power, PE: probability of exclusion,
  PIC:polymorphism information content
}
